requirements.txt
Django==6.0
PyJWT==2.10.1
PySocks==1.7.1
PyYAML==6.0.3
RapidFuzz==3.14.3
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
click==8.3.1
colorama==0.4.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
exceptiongroup==1.3.1
h11==0.16.0
idna==3.11
inflection==0.5.1
jsonschema-specifications==2025.9.1
jsonschema==4.25.1
kombu==5.6.1
numpy==2.2.6
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
pycparser==2.23
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
selenium==4.39.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.4
thefuzz==0.22.1
trio-websocket==0.12.2
trio==0.32.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.1
vine==5.1.0
wcwidth==0.2.14
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
